Updated Battle Formula
Changed battle formula, and some return methods to ensure that the
damage dealt was accurate, and that both the AI, and player used the
same formula.
AI works now that the player can die.
Need to fix the condition checking for when player dies.

+++ SDL/Moemon.h
@@ -48,8 +48,8 @@ public:
 	void setEnemy();
 	int getHealth() { return Health; };
 	int getLevel() { return Level; };
-	int getAttack() { return Attack; };
-	int getDefense() { return Defense; };
+	float getAttack() { return (float)Attack; };
+	float getDefense() { return (float)Defense; };
 	std::string getType() { return Type; };
 	std::string getName(){	return MoeMonName;	}
 	Text* getText() { if (Name == NULL) { return NULL; }  return Name; };
